+++ Sources/Sentry/SentrySDK.m
@@ -402,10 +402,7 @@ static NSUInteger startInvocations;
         }
     }
     [hub removeAllIntegrations];
-
-    // close the client
-    SentryClient *client = [hub getClient];
-    client.options.enabled = NO;
+    [hub close];
     [hub bindClient:nil];
 
     [SentrySDK setCurrentHub:nil];