+++ CHANGELOG.md
@@ -36,6 +36,7 @@ This version adds a dependency on Swift.
 - Marks App hang's event stacktrace snapshot as true (#2441) 
 - Enable user interaction tracing by default (#2442)
 - Remove default attachment content type (#2443)
+- SentrySDK.close calls flush (#2452)
 
 ## 7.31.2